Refactor integration test namespaces.
