+++ tests/integration/lib/Controllers/MetricExplorerTest.php
@@ -3,7 +3,7 @@
 namespace IntegrationTests\Controllers;
 
 use IntegrationTests\TokenAuthTest;
-use TestHarness\XdmodTestHelper;
+use IntegrationTests\TestHarness\XdmodTestHelper;
 
 class MetricExplorerTest extends TokenAuthTest
 {
