+++ tests/integration/lib/Rest/WarehouseControllerProviderTest.php
@@ -3,7 +3,7 @@
 namespace IntegrationTests\Rest;
 
 use IntegrationTests\TokenAuthTest;
-use TestHarness\XdmodTestHelper;
+use IntegrationTests\TestHarness\XdmodTestHelper;
 
 class WarehouseControllerProviderTest extends TokenAuthTest
 {
