+++ tests/integration/lib/Rest/WarehouseExportControllerProviderTest.php
@@ -9,7 +9,7 @@ use IntegrationTests\TokenAuthTest;
 use JsonSchema\Constraints\Constraint;
 use JsonSchema\Validator;
 use PHPUnit_Framework_TestCase;
-use TestHarness\XdmodTestHelper;
+use IntegrationTests\TestHarness\XdmodTestHelper;
 use XDUser;
 
 /**
