+++ tests/integration/lib/TestHarness/UserHelper.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TestHarness;
+namespace IntegrationTests\TestHarness;
 
 use XDUser;
