+++ tests/integration/lib/TestHarness/Utilities.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TestHarness;
+namespace IntegrationTests\TestHarness;
 
 use stdClass;
