+++ tests/unit/lib/OpenXdmod/Tests/HostListParserTest.php
@@ -3,7 +3,7 @@
  * @author Jeffrey T. Palmer <jtpalmer@buffalo.edu>
  */
 
-namespace OpenXdmod\Tests;
+namespace UnitTests\OpenXdmod\Tests;
 
 use Xdmod\HostListParser;
